requirements.txt
fastapi
uvicorn
python-multipart
passlib[bcrypt]
pyjwt